% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{Check}
\alias{Check}
\title{Run valgrind to find leaks in source code of a package}
\usage{
Check(
  pkgName,
  Archive = options()$refreshArchive,
  LogDir = options()$refreshCheckLog,
  lib = .libPaths()[1],
  recursive = FALSE,
  devel = FALSE,
  develR = "~/R/dev/R-devel/bin/R",
  ask = FALSE,
  as.cran = TRUE
)
}
\arguments{
\item{pkgName}{name of package}

\item{Archive}{place on computer where the tar.gz versions of the package live}

\item{LogDir}{place for output}

\item{lib}{R library}

\item{recursive}{if \code{TRUE} search also subdirectories of \code{Archive} for package versions}

\item{devel}{if \code{TRUE} use R's devel version}

\item{develR}{path to R's devel version}

\item{ask}{if \code{TRUE} let user choose a version}

\item{as.cran}{passed to R CMD check}
}
\description{
Run R CMD check on a version of a package
}
\details{
Run R CMD check on a version of a package
}
