\name{run}
\alias{run}
\title{Running examples}
\usage{
run(pkgName, Source = options()$refreshSource, valgrind = FALSE,
  verbose = 1)
}
\arguments{
  \item{pkgName}{}

  \item{Source}{}

  \item{valgrind}{}

  \item{verbose}{}
}
\value{
nothing
}
\description{
Running examples
}
\details{
This includes the \donttest{} sections of the examples.
}
\examples{
run(prodlim)
}
\author{
Thomas A. Gerds
}

