% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{Running test files}
\usage{
test(pkgName, Source = options()$packageHome, silent = 0L, verbose = 1)
}
\arguments{
\item{pkgName}{Name of package}

\item{Source}{directory in which the package source code lives}

\item{verbose}{jabber or not}

\item{valgrind}{does not work yet}
}
\value{
nothing
}
\description{
Running test files
}
\details{
Source all files in pkg/tests and run examples
}
\examples{

test(prodlim)

}
\author{
Thomas A. Gerds
}
