% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windoof.R
\name{windoof}
\alias{windoof}
\title{windoof}
\usage{
windoof(pgk, tarGet = "~/R/packages", recursive = FALSE)
}
\arguments{
\item{pgk}{package name}

\item{tarGet}{directory below which to find the tar.gz files of the package}

\item{recursive}{whether to search through subdirectories of tarGet}
}
\description{
Upload a package to winbuilder
}
