% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Build.R
\name{Build}
\alias{Build}
\title{R packaging tool combining emacs save buffers with
    devtools::document, devtools::build, devtools::install}
\usage{
Build(
  pkg,
  Source = options()$packageHome,
  lib = .libPaths()[1],
  ask = FALSE,
  quick = FALSE,
  recompile = FALSE,
  quiet = FALSE,
  vignettes = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{pkg}{The quoted or unquoted name of the package}

\item{Source}{A list of directories or a single directory #' in
    which to search for the package's source code. Default is
\code{options()$refreshSource}.}

\item{lib}{Where to install the package. Default is \code{.libPaths()[1]}.}

\item{ask}{If TRUE prompt user for source and library directories}

\item{quick}{passed to devtools::install}

\item{recompile}{use pkgbuild::compile_dll()?}

\item{quiet}{passed to devtools::install}

\item{vignettes}{if TRUE circumvent devtools::install and use R CMD build/INSTALL instead}

\item{verbose}{blabla?}

\item{...}{passed devtools::build}
}
\description{
Build an R-package from source files.
}
