% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Build.R
\name{Build}
\alias{Build}
\title{R packaging tool combining emacs save buffers with devtools::document, devtools::build, devtools::install}
\usage{
Build(pkg, Source = options()$packageHome, ask = FALSE,
  quick = FALSE, vignettes = FALSE, ...)
}
\arguments{
\item{pkg}{The quoted or unquoted name of the package}

\item{Source}{A list of directories or a single directory
in which to search for the package's source code. Default is \code{options()$refreshSource}.}

\item{ask}{If TRUE prompt user for source and library directories}

\item{vignettes}{passed to devtools::install as build_vignettes}

\item{...}{passed devtools::build}
}
\description{
Build an R-package from source files.
}
