% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Running examples}
\usage{
run(pkgName, Source = options()$refreshSource, valgrind = FALSE,
  verbose = 1)
}
\value{
nothing
}
\description{
Running examples
}
\details{
This includes the \donttest{} sections of the examples.
}
\examples{

run(prodlim)

}
\author{
Thomas A. Gerds
}

